+++ src/main/java/view/GameScreen.java
@@ -19,7 +19,8 @@ import java.io.InputStream;
 import java.util.ArrayList;
 
 /**
- * The type Game screen.
+ * The window from which the players participate in the SpaceRally.
+ *
  */
 public class GameScreen extends JPanel {
     private final ArrayList<CarShape> carShapes;
@@ -34,8 +35,8 @@ public class GameScreen extends JPanel {
     private JLabel timeLabel;
     private java.util.List<TypingPlayer> racers;
     private JPanel carPanel;
-    private java.util.List<ResultScreen> resultScreens = new ArrayList<>();
     private boolean isFinished = false;
+    private int wrongChars = 0;
 
     private long startTime;
     private int keyPressCount;
@@ -45,7 +46,6 @@ public class GameScreen extends JPanel {
     private SoundPlayer soundPlayer;
     private boolean soundOn;
     private SoundPlayer errorSoundPlayer;
-    private int wrongChars = 0;
 
     /**
      * Instantiates a new Game screen.
@@ -145,8 +145,9 @@ public class GameScreen extends JPanel {
                 int wpm = calculateWpm();
                 double accuracy = calculateAccuracy(typedText);
                 int progress = calculateProgress(typedText);
+
                 if (typedText.charAt(typedText.length() - 1) != providedText.charAt(typedText.length() - 1)
-                    && e.getKeyCode() != KeyEvent.VK_BACK_SPACE) {
+                        && e.getKeyCode() != KeyEvent.VK_BACK_SPACE) {
                     wrongChars++;
                 }
                 clientController.updateProgress(currentPlayer.getName(), wpm, progress, accuracy, timeElapsed);
@@ -203,12 +204,11 @@ public class GameScreen extends JPanel {
     }
 
     /**
-     * Add cars.
+     * Adds cars / UFOs and their racetracks to the game window.
      */
     public void addCars() {
         for (Player player : racers) {
             Car newCar = new Car(player);
-            //gameState.addPlayer(player);
             CarShape newCarShape = new CarShape(newCar, player,0, carShapes.size() * 50, 60, 50, customFont);
             carShapes.add(newCarShape);
             repaint();
@@ -235,15 +235,14 @@ public class GameScreen extends JPanel {
     }
 
     /**
-     * Update car positions.
-     *
-     * @param playerName the player name
-     * @param progress   the progress
-     * @param wpm        the wpm
+     * Updates the position of a specified player's car in the window.
+     * @param playerName the player to move
+     * @param progress the player's progress
+     * @param wpm the player's wpm count
      */
     public void updateCarPositions(String playerName, int progress, int wpm) {
         int totalLength = providedText.length();
-        double trackMultiplier = (isFinished) ? 0.915 : 0.7;
+        double trackMultiplier = (isFinished) ? 0.915 : 0.7; // To match the screen lengths
         int roadLength = (int) (carPanel.getWidth() * trackMultiplier); // 70% of the panel width
         for (CarShape carShape : carShapes) {
             if (carShape.getPlayer().getName().equals(playerName)) {
@@ -279,7 +278,7 @@ public class GameScreen extends JPanel {
                 correctChars++;
             }
         }
-        // Stopper for accuracy, this only really comes up when typing too fast
+        // Stopper for accuracy, this only really comes up when typing way too fast
         double accuracy = (correctChars * 100.0) / keyPressCount;
         if (accuracy > 100) return 100;
 
@@ -359,7 +358,7 @@ public class GameScreen extends JPanel {
             @Override
             protected void paintComponent(Graphics g) {
                 super.paintComponent(g);
-                double trackMultiplier = (isFinished) ? 0.9 : 0.7;
+                double trackMultiplier = (isFinished) ? 0.9 : 0.7; // To match the screen lengths
                 int roadLength = (int) (getWidth() * trackMultiplier); // 70% of the panel width
                 for (CarShape carShape : carShapes) {
                     carShape.draw(g, roadLength);